# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: POF API
  description: Productivity Orchestration Framework REST API
  version: 1.0.0
  contact:
    name: POF Development Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pof.example.com
    description: Production server
...
paths:
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [inbox, available, scheduled, blocked, completed, dropped]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, critical]
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: dueBefore
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    patch:
      tags: [Tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    delete:
      tags: [Tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
...
components:
  schemas:
    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [inbox, available, scheduled, blocked, completed, dropped]
        priority:
          type: string
          enum: [low, normal, high, critical]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
        dependsOn:
          type: array
          items:
            type: string
            format: uuid
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [inbox, available, scheduled, blocked, completed, dropped]
        priority:
          type: string
          enum: [low, normal, high, critical]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
        dependsOn:
          type: array
          items:
            type: string
            format: uuid
        tags:
          type: array
          items:
            type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [inbox, available, scheduled, blocked, completed, dropped]
        priority:
          type: string
          enum: [low, normal, high, critical]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items:
                  type: string
              message:
                type: string